refactor: move parser APIs to in-package
