+++ src/test/java/org/ingomohr/trac/TestTracMultiProtocolParser.java
@@ -6,6 +6,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 
+import org.ingomohr.trac.in.TracMultiProtocolParser;
 import org.ingomohr.trac.model.TracItem;
 import org.ingomohr.trac.model.TracProtocol;
 import org.ingomohr.trac.util.FileReader;
