+++ src/test/java/org/ingomohr/trac/examples/PrintTopicsByTimeSpent.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 
-import org.ingomohr.trac.TracProtocolParser;
+import org.ingomohr.trac.in.TracProtocolParser;
 import org.ingomohr.trac.model.TracProtocol;
 import org.ingomohr.trac.out.TracProtocolWriterByTopic;
 import org.ingomohr.trac.util.FileReader;
